Generate the file lists in the .wxs files.
Unfortunately, this is fairly hairy, due to the fact that .wxs file
lists are quite far from being simple lists of filenames.  Wix comes
with a tool (tallow) for creating these lists, but it seems to be
principally intended for creating an initial .wxs file, which you then
hand-edit.  So it's not a good fit for our needs here.  Instead, we
use find+awk to generate the File elements.