Execute consumer operation thunks in the provided task scheduler
*Submitting* consumer operations for execution takes a tiny
fraction of time but apparently some care enough about it to
request it to be performed on the provided scheduler.

This makes some logical sense. It can also lead to better
thread affinity or predictable throughput. But the primary
reason we do this is "if we don't, teh nerds will rage
again and again".

We don't want to anger the nerds, do we.

References #40, #41.