Rework heartbeat implementation
With this change .NET client uses the same approach as the Java
one.

Previously in 3.5.x we've used two timers that
re-scheduled themselves and used auto-reset events to be
notified about connection activity on the I/O loop.

This turned out to have some issues:

 * It was clever but fairly complicated
 * Picking the wrong heartbeat interval value led to sporadic
   failures due to race conditions [on the wire]
 * It didn't play well with socket timeouts

Our new strategy is more straightforward:
heartbeat writer uses a periodic timer that always sends
a frame, regardless of recent activity, every 1/2nd a timeout.

Given that heartbeat timeouts are often in minutes, this is not
unreasonable but possibly the most straightforward implementation
possible. It also doesn't suffer from race conditions on the wire
because we send a heartbeat frame every half the interval, so
slight inaccuracy in .NET runtime scheduling of timers won't
get in our way with reasonable (> 1 second) heartbeat intervals.

Heartbeat "reader" now relies on the socket receive timeout,
just like in the Java client. This is more straightforward
and there is no dissonance between I/O timeouts and the one we
use for incoming heartbeats.

Various cases when a SocketException could indicate a problem
other than a timeout (usually early in the connection
lifecycle) are accounted for.

References #68.