Switch heartbeat implementation back to latches (reset events)
We're back to square one with approaches to heartbeat implementations.
The socket-based solution has two issues:

 * Obscure TLS timeouts.
 * WinRT has incompatible I/O API which makes implementing the same
   socket timeout-based approach becomes a lot more challenging
   (a couple of days spent in an unsuccessful spike).

This reset event-driven approach had bugs in 3.5.0 but is quite
straightforward and I/O layer agnostic, so we're willing to go
back.

Note that we now have a significantly better test coverage in this
area compared to the 3.5.0 development cycle.