Locking on m_models in CreateModel and UnregisterModel.
RecoverModels method already locks on m_models collection. If CreateModel and UnregisterModel lock on a different object, then m_models collection can be modified while it is being read.
This is possible in concurrent scenarios, when the connection has been re-established, model recovery started, and then a new model created or disposed. In that case, there will be a "Collection was modified; enumeration operation may not execute" exception in RecoverModels and recovery will fail.