Fix stale delivery tags in RecoveryAwareModel
RecoveryAwareModel was attempting to prevent acking of stale delivery tags
by checking if the un-adjusted tag was > 0.  However, the tag is a ulong,
not a long, so it will always be >= 0.  This change compares the
un-adjusted tag to the adjusted tag to ensure it hasn't wrapped around to
a higher number.