Rename ReadFrame and WriteFrame
There are no "read frames" and "write frames" in the
protocol. "Incoming" and "outgoing" frames are also
not well established terms but at least they hint
at what the difference is and when they are used.

There are existing methods with identical names (ReadFrame, WriteFrame) as well,
which isn't helping.