RabbitMQ.Client.csproj: add support for targeting 4.6.2 framework
RabbitMQ added a new dependency on an obsolete Microsoft assembly
(Microsoft.Diagnostics.Tracing.EventSource 1.1.28). It is described as a "stop
gap" solution by Microsoft
(https://www.nuget.org/packages/Microsoft.Diagnostics.Tracing.EventSource) This
functionality is provided in the framework as of .NET 4.6, so providing a
version of the RabbitMQ dotnet client targeted to the 4.6.2 framework
(Microsoft's currently recommended framework) allows deployment of the RabbitMQ
dotnet client without having to redistribute the stop gap Microsoft assembly.

Signed-off-by: Keenan Forbes <keenan.forbes@ni.com>