Don't consider a heartbeat frame write failure to be fatal
Other clients do not do it [1], the spec doesn't instruct us to do so.
Let the heartbeat reader side detect peer unavailability (as it
currently does, using two missed heartbeats/intervals).

Heartbeat frame write can fail on on a highly congested connection
or a connection that's blocked by a resource alarm. The alarm might
clear soon enough for connection to be considered functional
by the peer.

Closes #638.

1. https://github.com/rabbitmq/rabbitmq-java-client/blob/5.6.x-stable/src/main/java/com/rabbitmq/client/impl/HeartbeatSender.java