Allow sending a data segment from a data buffer.
Sending a part of a data buffer may improve the client performances. The users of the API can now use a larger buffer for data serialization and then they don't have to copy out the segment of that data in a new byte array, and they can pass the buffer to the API indicating the position and the Length of tha data in buffer.