SslOption null protection
There are assumptions that the Ssl property will not be null. For example, ConnectionFactory.CreateConnection will fail to make an amqp connection if the Ssl property on the ConnectionFactory (which gets packged in an AmqpTcpEndpoint) is null due to an unprotected access of the Ssl.Enabled property when Ssl is null.