Re-introduce lock for all socket writes
The official `NetworkStream` docs are confusing:

https://docs.microsoft.com/en-us/dotnet/api/system.net.sockets.networkstream?view=netframework-4.5.1

> Read and write operations can be performed simultaneously on an
instance of the NetworkStream class without the need for
synchronization. As long as there is one unique thread for the write
operations and one unique thread for the read operations, there will be
no cross-interference between read and write threads and no
synchronization is required.

This is a poorly-written way to say "multiple threads must be synchronized".

Fixes #681