Fall back to TLSv1.2 when System.Security.Authentication.SslProtocols.None is disabled/unavailable
See #764 for background. In some environments we cannot assume
that SslProtocols.None will behave as expected:

 * It can be disabled via app context and other means
 * Its effective behavior is different between .NET versions according
   to .NET community blog posts and GitHub issues

So we borrow from [1] and special case this exception to retry with an
explicitly defined version (TLSv1.2, same as modern .NET default).

The discussion in [2] should also be mentioned as they have lead us
to understanding of many aspects of this intricate setting.

1. https://github.com/mysql-net/MySqlConnector/commit/715c3b542a1862b308f8afbb535ba1fee34afb93
2. https://github.com/robinrodricks/FluentFTP/issues/452#issuecomment-563965519