Made WorkService more responsive to stop signals under load
If the system is under load and has many tasks scheduled, the `Loop` can be scheduled but not started. Passing the cancellation token to `Task.Run` lets the `_worker` task that is being returned from the `Stop` call to be completed faster, so the consumer no longer has to wait for the thread pool to process the `Loop`. The synchronous continuation is added to ignore the cancellation exception in this scenario. The `Loop` code does not throw for now, but if it will, the continuation can be modified to re-throw the "inner" exceptions with `ExceptionDispatchInfo`.