Fold IAutorecoveringConnection [back] into IConnection
This undoes #758 (made in response to #515).

The events make more sense in the primary IConnection interface.
Most users enable automatic connection recovery and asking them to
cast to IAutorecoveringConnection is counter-intuitive.

As suggested in #515, we add no-op add and remove handlers
so that IConnection is easier to implement/mock.

See #940 for the discussion.

Closes #940.