fix: add unsigned short support in table deserialisation
While it seems that this wasn't needed in the lifetime of this library
so far, it seems weird not to support the full protocol as stated on the
RabbitMQ website (including errata).

This commit adds parsing support for the 'u' case, which should ensure
the library conforms fully to the protocol in those few edge cases where
it is still needed.

Inspired by: #662