Update Apigen to support .NET 6
Without specifically setting net6.0 as a target framework,
it is impossible to run Apigen project on M1 Macs, because
support for M1 Macs was added in .NET 6, according to this
post in MS forums [1].

To enable support for .NET 6, we had to add a NuGet dependency.
Everything else worked out of the box.

[1] https://learn.microsoft.com/en-us/answers/questions/793925/problem-installing-with-dotnet-on-m1-macos-archite

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>